+++ tests/bin/url-file/run.t
@@ -13,7 +13,7 @@ We need a basic opam project skeleton
     > opam-version: "2.0"\
     > homepage: "https://github.com/foo/whatever"\
     > dev-repo: "git+https://github.com/foo/whatever.git"\
-    > description: "whatever"\
+    > synopsis: "whatever"\
     > EOF
     $ touch README
     $ touch LICENSE
@@ -77,7 +77,7 @@ We make a dry-run release:
     [ OK ] File opam is present.
     -: exec: opam lint -s whatever.opam
     [ OK ] lint opam file whatever.opam.
-    [ OK ] opam field description is present
+    [ OK ] opam field synopsis is present
     [ OK ] opam fields homepage and dev-repo can be parsed by dune-release
     [ OK ] Skipping doc field linting, no doc field found
     [ OK ] lint of _build/whatever-0.1.0 and package whatever success
@@ -134,9 +134,8 @@ We make a dry-run release:
     opam-version: "2.0"
     homepage: "https://github.com/foo/whatever"
     dev-repo: "git+https://github.com/foo/whatever.git"
-    description: "whatever"
+    synopsis: "whatever"
     x-commit-hash: ...
-    synopsis: ""
     url {
       src: "https://foo.fr/archive/foo/foo.tbz"
       checksum: [